Check correct connection header for client
