+++ ember-server/src/main/scala/org/http4s/ember/server/internal/ServerHelpers.scala
@@ -17,7 +17,6 @@
 package org.http4s.ember.server.internal
 
 import cats._
-import cats.data.NonEmptyList
 import cats.effect._
 import cats.effect.concurrent._
 import cats.effect.implicits._
@@ -27,26 +26,19 @@ import fs2.Stream
 import fs2.io.tcp._
 import fs2.io.tls._
 import java.net.InetSocketAddress
-import java.util.Locale
 import org.http4s._
-import org.http4s.ember.core.Util.durationToFinite
+import org.http4s.ember.core.Util._
 import org.http4s.ember.core.{Drain, EmptyStreamError, Encoder, Parser, Read}
-import org.http4s.headers.{Connection, Date}
+import org.http4s.headers.Date
 import org.http4s.internal.tls.{deduceKeyLength, getCertChain}
 import org.http4s.server.{SecureSession, ServerRequestKeys}
-import org.typelevel.ci._
 import org.typelevel.log4cats.Logger
 import org.typelevel.vault.Vault
 import scala.concurrent.duration._
 import scodec.bits.ByteVector
+import org.http4s.headers.Connection
 
 private[server] object ServerHelpers {
-
-  private[this] val closeCi = ci"close"
-  private[this] val keepAliveCi = ci"keep-alive"
-  private[this] val connectionCi = ci"connection"
-  private[this] val close = Connection(NonEmptyList.of(closeCi))
-  private[this] val keepAlive = Connection(NonEmptyList.one(keepAliveCi))
 
   private val serverFailure =
     Response(Status.InternalServerError).putHeaders(org.http4s.headers.`Content-Length`.zero)
@@ -175,30 +167,10 @@ private[server] object ServerHelpers {
   private[internal] def postProcessResponse[F[_]: Timer: Monad](
       req: Request[F],
       resp: Response[F]): F[Response[F]] = {
-    val connection: Connection =
-      if (isKeepAlive(req.httpVersion, req.headers)) keepAlive
-      else close
+    val connection = connectionFor(req.httpVersion, req.headers)
     for {
       date <- HttpDate.current[F].map(Date(_))
     } yield resp.withHeaders(Headers(date, connection) ++ resp.headers)
-  }
-
-  private[internal] def isKeepAlive(httpVersion: HttpVersion, headers: Headers): Boolean = {
-    // We know this is raw because we have not parsed any headers in the underlying alg.
-    // If Headers are being parsed into processed for in ParseHeaders this is incorrect.
-    // TODO: the problem is that any string that contains `expected` is admissible
-    def hasConnection(expected: String): Boolean =
-      headers.headers.exists {
-        case Header.Raw(name, value) =>
-          name == connectionCi && value.toLowerCase(Locale.ROOT).contains(expected)
-        case _ => false
-      }
-
-    httpVersion match {
-      case HttpVersion.`HTTP/1.0` => hasConnection(keepAliveCi.toString)
-      case HttpVersion.`HTTP/1.1` => !hasConnection(closeCi.toString)
-      case _ => false
-    }
   }
 
   private[internal] def runConnection[F[_]: Concurrent: Timer](
